requirements.txt
asgiref==3.2.3
Django==3.0.3
django-model-utils==4.0.0
django-multiselectfield==0.1.11
django-select-multiple-field==0.4.2
pytz==2019.3
sqlparse==0.3.0
